fix(money): move BalanceProjection into Money components folder (MUSD-945)
