+++ app/components/UI/Money/components/BalanceProjection/index.ts
@@ -1,0 +1,2 @@
+export { BalanceProjection } from './BalanceProjection';
+export type { BalanceProjectionProps } from './BalanceProjection';
