--- app/components/Views/confirmations/components/balance-projection/index.ts
@@ -1,2 +1,0 @@
-export { BalanceProjection } from './balance-projection';
-export type { BalanceProjectionProps } from './balance-projection';
